scripts/requirements.txt
ruamel.yaml
python-frontmatter
Jinja2
markdown-it-py
pytest
beautifulsoup4
jsonschema
